+++ src/compiler/codegen_c_json_parser.c
@@ -416,6 +416,8 @@ static int gen_field_match_handler(fb_output_t *out, fb_compound_type_t *ct, voi
     int is_base64 = 0;
     int is_base64url = 0;
     int is_nested = 0;
+    int is_array = 0;
+    size_t array_len = 0;
     int st = 0;
     const char *tname_prefix = "n/a", *tname = "n/a"; /* suppress compiler warnigns */
 
@@ -454,7 +456,15 @@ static int gen_field_match_handler(fb_output_t *out, fb_compound_type_t *ct, voi
         is_nested = member->nest != 0;
         is_base64 = member->metadata_flags & fb_f_base64;
         is_base64url = member->metadata_flags & fb_f_base64url;
-        /* Fall through. */
+        is_scalar = 1;
+        st = member->type.st;
+        break;
+    case vt_fixed_array_type:
+        is_array = 1;
+        is_scalar = 1;
+        array_len = member->type.len;
+        st = member->type.st;
+        break;
     case vt_scalar_type:
         is_scalar = 1;
         st = member->type.st;
@@ -515,6 +525,13 @@ repeat_nested:
                 (uint64_t)member->size, (short)member->align,
                 (uint64_t)FLATBUFFERS_COUNT_MAX(member->size));
         }
+    }
+    if (is_array) {
+        println(out, "size_t count = %d;", array_len);
+        println(out, "%s *base = (%s *)((size_t)struct_base + %"PRIu64");",
+                tname, tname, (uint64_t)member->offset);
+    }
+    if (is_array || is_vector) {
         println(out, "buf = flatcc_json_parser_array_start(ctx, buf, end, &more);");
         /* Note that we reuse `more` which is safe because it is updated at the end of the main loop. */
         println(out, "while (more) {"); indent();
@@ -542,7 +559,7 @@ repeat_nested:
     if (is_vector && !is_offset) {
         println(out, "if (!(pval = flatcc_builder_extend_vector(ctx->ctx, 1))) goto failed;");
     }
-    if (is_struct_container) {
+    if (is_struct_container && !is_array) {
         /* `struct_base` is given as argument to struct parsers. */
         println(out, "pval = (void *)((size_t)struct_base + %"PRIu64");", (uint64_t)member->offset);
     } else if (is_struct && !is_vector) {
@@ -560,7 +577,7 @@ repeat_nested:
         if (!is_struct_container && !is_vector && !is_base64 && !is_base64url) {
 #if !FLATCC_JSON_PARSE_FORCE_DEFAULTS
             /* We need to create a check for the default value and create a table field if not the default. */
-            switch(member->value.type) {
+            switch (member->value.type) {
             case vt_bool:
             case vt_uint:
                 println(out, "if (val != %"PRIu64" || (ctx->flags & flatcc_json_parser_f_force_add)) {", member->value.u); indent();
@@ -597,7 +614,16 @@ repeat_nested:
 #endif
         }
         /* For scalars in table field, and in struct container. */
-        println(out, "%s%s_write_to_pe(pval, val);", out->nsc, tname_prefix);
+        if (is_array) {
+            println(out, "if (count) {"); indent();
+            println(out, "%s%s_write_to_pe(base, val);", out->nsc, tname_prefix);
+            println(out, "--count, ++base;");
+            unindent(); println(out, "} else if (!(ctx->flags & flatcc_json_parser_f_skip_array_overflow)) {"); indent();
+            println(out, "return flatcc_json_parser_set_error(ctx, buf, end, flatcc_json_parser_error_array_overflow);");
+            unindent(); println(out, "}");
+        } else {
+            println(out, "%s%s_write_to_pe(pval, val);", out->nsc, tname_prefix);
+        }
         if (!is_struct_container && !is_vector) {
             unindent(); println(out, "}");
         }
@@ -650,6 +676,16 @@ repeat_nested:
         } else {
             println(out, "ref = flatcc_builder_end_vector(ctx->ctx);");
         }
+    }
+    if (is_array) {
+        println(out, "buf = flatcc_json_parser_array_end(ctx, buf, end, &more);");
+        unindent(); println(out, "}");
+        println(out, "if (count) {"); indent();
+        println(out, "if (!(ctx->flags & flatcc_json_parser_f_pad_array_underflow)) {"); indent();
+        println(out, "return flatcc_json_parser_set_error(ctx, buf, end, flatcc_json_parser_error_array_underflow);");
+        unindent(); println(out, "}");
+        println(out, "memset(base, 0, count * sizeof(*base));");
+        unindent(); println(out, "}");
     }
     if (is_nested == 1) {
         is_nested = 2;